# docs/API_Contract_260714.yaml
openapi: 3.1.0
info:
  title: HUG 안심전세 체인 API Contract
  version: "1.0.0"
  summary: >-
    Frontend / Backend / ML / Blockchain 공통 계약(Single Source of Truth) 문서.
    Backend 구현 문서가 아니며 FastAPI 코드도 아니다.
  description: |
    ## 문서 성격
    본 문서는 HUG × 아이엔 안심전세 체인(HUG 안심전세 체인) 프로젝트의 **Frontend / Backend / ML / Blockchain 공통 계약**이다.
    Backend를 새로 설계하는 문서가 아니라, 이미 확정된 4개 상위 문서(개발설계보고서, 데이터수집 및 API가이드, ML개발가이드,
    Frontend UI/UX 명세서)를 근거로 프론트엔드가 axios 호출을 만들고, 백엔드가 FastAPI를 구현하고, ML Notebook이
    inference endpoint를 구현할 수 있도록 하는 **OpenAPI 3.1 단일 기준선**이다.

    ## 서비스 한 줄 요약
    HUG 안심전세 체인은 위험을 점수화하는 앱이 아니라, 계약 전 위험을 발견하고 위험조건이 실제로 해소되었는지
    증빙과 외부 API로 검증하며, 계약서·검증 이력을 블록체인에 공증하고, 계약 만료 전 D-90 반환계획 확인과
    사고 후 HUG 채권관리 인계까지 하나의 생애주기로 연결하는 주거안전 인프라다.

    ## 선행 문서와 우선순위
    | 영역 | 기준 문서 | 충돌 시 우선순위 |
    |---|---|---|
    | 공통 식별자·상태 | 개발설계보고서_260714_수정보완.docx 10장, 부록 B·C | 최우선 |
    | 외부 API/Mock | 데이터수집_및_API가이드_260714.md | 외부 API 상세 |
    | ML 입출력 계약 | ML개발가이드_260714.md 20장, 부록 C | 모델 Request/Response |
    | 화면·Route·API Mapping | Frontend_UIUX_명세서_260714.md 2·5·8장 | 화면-API 연결 |

    본 문서(API_Contract_260714.yaml)는 위 4개 문서의 API 관련 내용을 통합한 **실행 가능한 계약**이며,
    향후 작성될 Backend_API_명세서_260714.md, Blockchain_설계서_260714.md는 본 계약과 충돌할 수 없다.

    ## 데이터 신뢰성에 대한 명시 (ML개발가이드 2·30장 계승)
    - 정상계약(사고 없이 종료된 계약) 데이터가 없어 "실제 사고확률"을 추정하는 모델은 존재하지 않는다.
    - HUG 사고·경매·배당 데이터 7종은 전량 합성데이터이며, 회수·처리기간 관련 응답은 "합성데이터 기준 패턴"이다.
...

    ## 온체인 금지 원칙 (개발설계보고서 8장)
...

    ## 문서 요약 통계
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [Contract]
      summary: 계약 목록 조회
      description: 로그인 사용자(tenant/landlord) 기준 진행 중인 계약 목록을 조회한다. TEN-00/LAND-00 홈 화면에서 사용한다.
      x-permission: tenant, landlord
      x-mock-supported: true
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: contract_status
          in: query
          required: false
          description: ContractStatus 필터
          schema:
            $ref: '#/components/schemas/ContractStatus'
      responses:
        "200":
          description: 계약 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Contract'
                          pagination:
                            $ref: '#/components/schemas/Pagination'
              example:
                status: Success
                data:
                  items:
                    - contract_id: "f4a5b6c7-0000-0000-0000-000000000040"
                      property_id: "c1a2b3c4-0000-0000-0000-000000000010"
                      contract_status: Monitoring
                      deposit: 400000000
                      contract_start_date: "2026-08-01"
                      contract_end_date: "2028-07-31"
                  pagination: { page: 1, size: 20, total_elements: 1, total_pages: 1 }
                request_id: "req_20260714_0008"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createContract
      tags: [Contract]
      summary: 계약 생성
      description: |
        주소·보증금·계약기간·임대인 유형을 입력받아 contract_id를 생성한다(개발설계보고서 MVP 완료 기준, 표 11).
        생성 직후 ContractStatus는 Draft이며, 이후 위험진단 완료 시 Diagnosed로 전이한다(전이 상세는 Backend_API_명세서에서 확정).
      x-permission: tenant
      x-mock-supported: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractCreateRequest'
            example:
              property_id: "c1a2b3c4-0000-0000-0000-000000000010"
              deposit: 400000000
              contract_start_date: "2026-08-01"
              contract_end_date: "2028-07-31"
              landlord_type: INDIVIDUAL
              housing_type: MULTI_HOUSEHOLD
      responses:
        "201":
          description: 계약 생성 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Contract'
              example:
                status: Success
                data:
                  contract_id: "f4a5b6c7-0000-0000-0000-000000000040"
                  property_id: "c1a2b3c4-0000-0000-0000-000000000010"
                  contract_status: Draft
                  deposit: 400000000
                  contract_start_date: "2026-08-01"
                  contract_end_date: "2028-07-31"
                  created_at: "2026-07-14T13:00:00+09:00"
                request_id: "req_20260714_0009"
        "422":
          $ref: '#/components/responses/ValidationError'
        "409":
          $ref: '#/components/responses/Conflict'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /contracts/{contract_id}:
    get:
      operationId: getContract
      tags: [Contract]
      summary: 계약 상세 조회
      description: contract_id 기준 계약 상세(물건 정보, 현재 ContractStatus, 최신 버전 참조)를 조회한다.
      x-permission: tenant, landlord (본인 계약만), advisor, hug_admin, system_admin (읽기 전용)
      x-mock-supported: true
      parameters:
        - $ref: '#/components/parameters/ContractIdPath'
      responses:
        "200":
          description: 계약 상세 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Contract'
        "404":
          $ref: '#/components/responses/NotFound'
        "403":
          $ref: '#/components/responses/Forbidden'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /contracts/{contract_id}/timeline:
    get:
      operationId: getContractTimeline
      tags: [Contract]
      summary: 계약 타임라인 조회
      description: |
        계약 생애주기 상태 변경 이벤트(TimelineEvent, 10.3절 공통 이벤트)와 각 이벤트의 BlockchainStatus를
        시간순으로 조회한다. TEN-05, HUG-02 화면에서 사용한다.
      x-permission: tenant, landlord (본인 계약만), advisor, hug_admin
      x-mock-supported: true
      parameters:
        - $ref: '#/components/parameters/ContractIdPath'
      responses:
        "200":
          description: 타임라인 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          contract_id:
                            $ref: '#/components/schemas/UUID'
                          contract_status:
                            $ref: '#/components/schemas/ContractStatus'
                          events:
                            type: array
                            items:
                              $ref: '#/components/schemas/TimelineEvent'
              example:
                status: Success
                data:
                  contract_id: "f4a5b6c7-0000-0000-0000-000000000040"
                  contract_status: Monitoring
                  events:
                    - timeline_event_id: "aa000000-0000-0000-0000-000000000001"
                      event_type: ContractCreated
                      occurred_at: "2026-07-01T09:00:00+09:00"
                      blockchain_status: Confirmed
                    - timeline_event_id: "aa000000-0000-0000-0000-000000000002"
                      event_type: RiskAssessed
                      occurred_at: "2026-07-02T10:00:00+09:00"
                      blockchain_status: Confirmed
                    - timeline_event_id: "aa000000-0000-0000-0000-000000000003"
                      event_type: EvidenceSubmitted
                      occurred_at: "2026-07-05T11:00:00+09:00"
                      blockchain_status: Pending
                request_id: "req_20260714_0010"
        "404":
          $ref: '#/components/responses/NotFound'
        "403":
          $ref: '#/components/responses/Forbidden'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /contracts/{contract_id}/return-plan:
    get:
      operationId: getReturnPlan
      tags: [Contract]
      summary: D-90 반환계획 조회
      description: |
        계약 만료 90일 전 반환 준비 상태(ReturnPlan), D-90 카운트다운, 임대인 응답 여부, 조기경보 배지를 조회한다.
        TEN-06 화면에서 사용하며, 반환계획 자체의 제출은 POST /return-plans(임대인)에서 수행한다.
      x-permission: tenant, landlord (본인 계약만)
      x-mock-supported: true
      parameters:
        - $ref: '#/components/parameters/ContractIdPath'
      responses:
        "200":
          description: 반환계획 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReturnPlan'
              example:
                status: Success
                data:
                  return_plan_id: "bb000000-0000-0000-0000-000000000001"
                  contract_id: "f4a5b6c7-0000-0000-0000-000000000040"
                  d_day: -62
                  landlord_response_status: NotResponded
                  early_warning: true
                  planned_return_date: null
                  created_at: "2026-07-14T00:00:00+09:00"
                request_id: "req_20260714_0011"
        "404":
          $ref: '#/components/responses/NotFound'
        "403":
          $ref: '#/components/responses/Forbidden'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
  /return-plans:
    post:
      operationId: submitReturnPlan
      tags: [Evidence-Verification]
      summary: 반환계획 제출
      description: D-90 도달 시 임대인이 보증금 반환 예정일·방법을 제출한다. LAND-03 화면에서 호출한다.
      x-permission: landlord
      x-mock-supported: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnPlanCreateRequest'
            example:
              contract_id: "f4a5b6c7-0000-0000-0000-000000000040"
              planned_return_date: "2028-07-31"
              return_method: "신규 임차인 보증금으로 반환"
              note: "특이사항 없음"
      responses:
        "201":
          description: 반환계획 제출 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReturnPlan'
              example:
                status: Success
                data:
                  return_plan_id: "bb000000-0000-0000-0000-000000000001"
                  contract_id: "f4a5b6c7-0000-0000-0000-000000000040"
                  landlord_response_status: Responded
                  planned_return_date: "2028-07-31"
                request_id: "req_20260714_0016"
        "422":
          $ref: '#/components/responses/ValidationError'
        "409":
          $ref: '#/components/responses/Conflict'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer 토큰. `role` 클레임(tenant/landlord/advisor/hug_admin/system_admin/verifier)을 포함한다.
        `verifier`는 advisor 내 서브 권한으로, `/verifications/{id}/decision` 액션 권한 판정에만 사용한다
        (Frontend_UIUX_명세서 3장, 개발설계보고서 13.5절). 발급/폐기 상세 구현은 Backend_API_명세서_260714.md에서 확정한다.
  parameters:
...
    ContractIdPath:
      name: contract_id
      in: path
      required: true
      description: 계약 식별자
      schema:
        $ref: '#/components/schemas/UUID'
...
    PageParam:
      name: page
      in: query
      required: false
      description: 페이지 번호(1부터 시작)
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      required: false
      description: 페이지당 항목 수
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    ValidationError:
      description: 요청 값 검증 실패(422). Zod/Pydantic 검증 오류를 필드 단위로 반환한다.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      error_code:
                        type: string
                        example: ERROR-002
                      message:
                        type: string
                        example: "요청 값이 유효하지 않습니다."
                      field_errors:
                        type: array
                        items:
                          $ref: '#/components/schemas/ValidationErrorItem'
          example:
            status: Failed
            error:
              error_code: ERROR-002
              message: "요청 값이 유효하지 않습니다."
              field_errors:
                - field: deposit
                  reason: "0보다 큰 정수여야 합니다."
            request_id: "req_20260714_0100"
    Unauthorized:
      description: 인증 실패(401). 토큰 없음/만료/서명 불일치.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-004, message: "인증 토큰이 유효하지 않습니다." }
            request_id: "req_20260714_0101"
    Forbidden:
      description: 권한 없음(403). 역할(role)이 허용되지 않거나 verifier 서브 권한이 없는 경우.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-005, message: "해당 작업을 수행할 권한이 없습니다." }
            request_id: "req_20260714_0102"
    NotFound:
      description: 리소스 없음(404). 존재하지 않거나 본인 소유가 아닌 리소스에 대한 조회.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-006, message: "요청한 리소스를 찾을 수 없습니다." }
            request_id: "req_20260714_0103"
    Conflict:
      description: 상태 충돌(409). 중복 접수, 상태 전이 규칙 위반 등.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-007, message: "이미 접수된 사건이 존재합니다." }
            request_id: "req_20260714_0104"
...
  schemas:
...
    UUID:
      type: string
      format: uuid
      description: 전역 고유 식별자. 개발설계보고서 10.1절 공통 식별자 16종은 모두 이 포맷을 사용한다.
      example: "8f14e45f-ceea-4b4f-9c1a-000000000001"
    Timestamp:
      type: string
      format: date-time
      description: KST 기준 ISO 8601 일시(개발설계보고서 10.5절). 오프셋 `+09:00`을 포함해 표기한다.
      example: "2026-07-14T13:30:00+09:00"
    Money:
      type: integer
      format: int64
      minimum: 0
      description: KRW 정수 원 단위 금액(개발설계보고서 10.5절). 화면에서는 천 단위 구분을 적용해 표시한다.
      example: 400000000
    Pagination:
      type: object
      description: 목록 API 공통 페이지네이션 응답.
      required: [page, size, total_elements, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        size:
          type: integer
          minimum: 1
          example: 20
        total_elements:
          type: integer
          minimum: 0
          example: 1
        total_pages:
          type: integer
          minimum: 0
          example: 1
...
    SuccessResponse:
      type: object
      description: 모든 성공 응답의 공통 봉투(envelope). data 필드는 각 API별 실제 스키마로 대체된다.
      required: [status, data, request_id]
      properties:
        status:
          type: string
          enum: [Success, Partial]
          example: Success
        data:
          description: 실제 응답 페이로드. 각 API operation에서 구체 스키마로 재정의한다.
          type: object
        request_id:
          type: string
          description: Request-ID 헤더와 동일한 값을 본문에도 포함해 로그 상관관계를 보장한다.
          example: "req_20260714_0001"
    ErrorResponse:
      type: object
      description: 모든 오류 응답의 공통 봉투. error_code는 "ERROR-NNN" 형식(하단 오류코드 표 참조).
      required: [status, error, request_id]
      properties:
        status:
          type: string
          enum: [Failed]
          example: Failed
        error:
          type: object
          required: [error_code, message]
          properties:
            error_code:
              type: string
              pattern: '^ERROR-[0-9]{3}$'
              example: ERROR-006
            message:
              type: string
              example: "요청한 리소스를 찾을 수 없습니다."
            details:
              type: [object, "null"]
              additionalProperties: true
        request_id:
          type: string
          example: "req_20260714_9999"
    ValidationErrorItem:
      type: object
      description: 필드 단위 검증 오류 항목.
      required: [field, reason]
      properties:
        field:
          type: string
          example: deposit
        reason:
          type: string
          example: "0보다 큰 정수여야 합니다."
    # ======================================================================
    # 2. 공통 Enum (Common Enum) - 14종 (개발설계보고서 10.2절, 부록 C 및 관련 문서 기준)
    # ======================================================================
    ContractStatus:
      type: string
      description: 계약 생애주기 상태(개발설계보고서 10.2절, 8단계 생애주기와 1:1 대응). 전이 상세는 Backend_API_명세서_260714.md에서 확정한다.
      enum:
        - Draft
        - Diagnosed
        - EvidenceRequested
        - EvidenceSubmitted
        - Verified
        - ContractFinalized
        - Monitoring
        - D90Requested
        - ReturnPlanSubmitted
        - AtRisk
        - IncidentReported
        - TransferredToHUG
        - RecoveryInProgress
        - Closed
      example: Monitoring
...
    BlockchainStatus:
      type: string
      description: 온체인 기록 상태(개발설계보고서 10.2절). Confirmed라도 법적 효력을 의미하지 않는다(Frontend_UIUX_명세서 10.4절).
      enum: [NotRequested, Pending, Confirmed, Failed]
      example: Confirmed
...
    HousingType:
      type: string
      description: 주택유형(ML개발가이드 10장 Feature Engineering, 데이터수집_및_API가이드 건축물대장 연계).
      enum:
        - MULTI_HOUSEHOLD
        - MULTI_FAMILY
        - APARTMENT
        - OFFICETEL
        - SINGLE_FAMILY
        - ROW_HOUSE
        - OTHER
      example: MULTI_HOUSEHOLD
    LandlordType:
      type: string
      description: 임대인 유형(개발설계보고서 표 9, ML개발가이드 부록 B landlord_type Feature).
      enum: [INDIVIDUAL, INDIVIDUAL_BUSINESS, CORPORATION]
      example: INDIVIDUAL
    # ======================================================================
    # 3. Entity Schema - 19종 (개발설계보고서 13.3절 핵심 엔터티 + Frontend/HUG 화면에 필요한 보조 엔터티)
    # ======================================================================
...
    Contract:
      type: object
      description: 계약 생애주기 중심 엔터티(개발설계보고서 부록 B contract_id, 13.4절 ERD).
      required: [contract_id, property_id, contract_status, deposit]
      properties:
        contract_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        property_id:
          $ref: '#/components/schemas/UUID'
        tenant_user_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        landlord_user_id:
          type: [string, "null"]
          format: uuid
          readOnly: true
        contract_status:
          $ref: '#/components/schemas/ContractStatus'
        deposit:
          $ref: '#/components/schemas/Money'
        contract_start_date:
          type: string
          format: date
          example: "2026-08-01"
        contract_end_date:
          type: string
          format: date
          example: "2028-07-31"
        landlord_type:
          $ref: '#/components/schemas/LandlordType'
        housing_type:
          $ref: '#/components/schemas/HousingType'
        latest_contract_version_id:
          type: [string, "null"]
          format: uuid
          readOnly: true
        created_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
        updated_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
...
    TimelineEvent:
      type: object
      description: 계약 생애주기 이벤트(개발설계보고서 부록 B timeline_event_id, 10.3절 공통 이벤트 13종).
      required: [timeline_event_id, contract_id, event_type, occurred_at]
      properties:
        timeline_event_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        contract_id:
          $ref: '#/components/schemas/UUID'
        event_type:
          type: string
          enum:
            - ContractCreated
            - RegistrySnapshotCreated
            - RiskAssessed
            - EvidenceRequested
            - EvidenceSubmitted
            - VerificationCompleted
            - ContractVersionFinalized
            - D90Requested
            - ReturnPlanSubmitted
            - IncidentReported
            - TransferredToHUG
            - RecoveryPredictionCreated
            - BlockchainConfirmed
          example: RiskAssessed
        occurred_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
        blockchain_status:
          $ref: '#/components/schemas/BlockchainStatus'
        blockchain_tx_id:
          type: [string, "null"]
          format: uuid
    ReturnPlan:
      type: object
      description: D-90 반환계획 엔터티(개발설계보고서 부록 B return_plan_id).
      required: [return_plan_id, contract_id]
      properties:
        return_plan_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        contract_id:
          $ref: '#/components/schemas/UUID'
        d_day:
          type: integer
          description: 계약 만료일 기준 D-day(음수=만료 전 잔여일).
          example: -62
        landlord_response_status:
          type: string
          enum: [NotResponded, Responded, ReturnDifficultyReported]
          example: NotResponded
        early_warning:
          type: boolean
          default: false
        planned_return_date:
          type: [string, "null"]
          format: date
        return_method:
          type: [string, "null"]
        note:
          type: [string, "null"]
        blockchain_tx_id:
          type: [string, "null"]
          format: uuid
          readOnly: true
        created_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
...
    ContractCreateRequest:
      type: object
      required: [property_id, deposit, contract_start_date, contract_end_date, landlord_type, housing_type]
      properties:
        property_id:
          $ref: '#/components/schemas/UUID'
        deposit:
          $ref: '#/components/schemas/Money'
        contract_start_date:
          type: string
          format: date
        contract_end_date:
          type: string
          format: date
        landlord_type:
          $ref: '#/components/schemas/LandlordType'
        housing_type:
          $ref: '#/components/schemas/HousingType'
...
    ReturnPlanCreateRequest:
      type: object
      required: [contract_id, planned_return_date, return_method]
      properties:
        contract_id:
          $ref: '#/components/schemas/UUID'
        planned_return_date:
          type: string
          format: date
        return_method:
          type: string
          example: "신규 임차인 보증금으로 반환"
        note:
          type: [string, "null"]
...